refact: made code more modular
	- completed some todos and removed others

	- made come more modular by removing reusable components from parent component

	- implemented a date formater helper function
