+++ package.json
@@ -11,6 +11,7 @@
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "chart.js": "^4.4.7",
+    "date-fns": "^4.1.0",
     "motion": "^11.12.0",
     "react": "^18.3.1",
     "react-chartjs-2": "^5.2.0",
