+++ src/app/theme/app_colors.tsx
@@ -1,5 +1,3 @@
-// TODO: add hard coded app colors to AppColors class with their associated name
-
 class AppColors {
   static grey300 = "#B3B3B3";
 }
