+++ src/features/auth/login/login_view.tsx
@@ -1,7 +1,5 @@
 import AuthForm from "../shared/components/auth_form";
 
-// TODO: fix Login View Spacing
-
 const LoginView = (): JSX.Element => {
   return (
     <>
