+++ src/features/auth/shared/css/auth_form.css
@@ -10,7 +10,6 @@
     display: flex;
     flex-direction: column;
     width: 560px;
-    height: 422px;
     background-color: white;
     border-radius: 16px;
     padding: 32px;
