+++ src/features/shared/components/css/text_form_field.css
@@ -1,5 +1,3 @@
-/*!! TODO: refactor css order and conver px to rem */
-
 .text-form-field-container {
     display: flex;
     width: 100%;
