+++ src/features/shared/hooks/use_selectable_list_tile_listeners.tsx
@@ -1,8 +1,6 @@
 import { useEffect } from "react";
 import { useLocation } from "react-router";
 
-// TODO: use some form of toLower() so it doesnt matter if the route is upper of lowercase
-
 // TODO: will need to be updated when adding other routes with spaces | maybe parse element.textContent
 
 const useAddSelectableListTileListeners = ({
@@ -19,7 +17,11 @@ const useAddSelectableListTileListeners = ({
     const elements = document.querySelectorAll(selector);
 
     elements.forEach((element) => {
-      if (location.pathname.includes(element.textContent!)) {
+      if (
+        location.pathname
+          .toLowerCase()
+          .includes(element.textContent!.toLowerCase())
+      ) {
         element.classList.add(selectedStyle);
       } else {
         element.classList.remove(selectedStyle);
